+++ packages/elements-web-components/src/__stories__/Api.ts
@@ -1,10 +1,25 @@
 import '../index';
 
+import { zoomApiYaml } from '@stoplight/elements/src/__fixtures__/api-descriptions/zoomApiYaml';
+import { parse } from '@stoplight/yaml';
+
 import { createTemplate } from './util';
 
 const Template = createTemplate('elements-api');
 
-export default { title: 'API' };
+export default {
+  title: 'API',
+  argTypes: {
+    apiDescriptionUrl: {
+      control: 'text',
+    },
+    apiDescriptionDocument: { control: 'text' },
+    layout: {
+      control: { type: 'inline-radio', options: ['sidebar', 'stacked'] },
+      defaultValue: 'sidebar',
+    },
+  },
+};
 
 export const TodosAPI = Template.bind({});
 TodosAPI.args = {
@@ -16,3 +31,15 @@ ZoomApi.args = {
   apiDescriptionUrl: 'https://raw.githubusercontent.com/stoplightio/Public-APIs/master/reference/zoom/zoom.yaml',
 };
 ZoomApi.storyName = 'Complex API with inline `$ref`s';
+
+export const APIWithYamlProvidedDirectly = Template.bind({});
+APIWithYamlProvidedDirectly.args = {
+  apiDescriptionDocument: zoomApiYaml,
+};
+APIWithYamlProvidedDirectly.storyName = 'API With Yaml Provided Directly';
+
+export const APIWithJSONProvidedDirectly = Template.bind({});
+APIWithJSONProvidedDirectly.args = {
+  apiDescriptionDocument: JSON.stringify(parse(zoomApiYaml), null, '  '),
+};
+APIWithJSONProvidedDirectly.storyName = 'API With JSON Provided Directly';