+++ CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [6.0.0-beta.171](https://github.com/stoplightio/elements/compare/v6.0.0-beta.170...v6.0.0-beta.171) (2020-10-23)
+
+
+### chore
+
+* remove `linkComponent` input from API/StoplightProject ([#665](https://github.com/stoplightio/elements/issues/665)) ([3bc208d](https://github.com/stoplightio/elements/commit/3bc208de413061d7057428e85d1666976832ac0c))
+
+
+### BREAKING CHANGES
+
+* removed prop
+
+* test: add storybook controls to ease testing
+
+
+
+
+
+# [6.0.0-beta.170](https://github.com/stoplightio/elements/compare/v6.0.0-beta.169...v6.0.0-beta.170) (2020-10-19)
+
+
+### Bug Fixes
+
+* **deps:** [security] bump npm-user-validate from 1.0.0 to 1.0.1 ([#630](https://github.com/stoplightio/elements/issues/630)) ([5964268](https://github.com/stoplightio/elements/commit/5964268cbfdeb291e43695f400aafa1abf56df79))
+
+
+### Features
+
+* allow to provide api document string directly in API web component ([#629](https://github.com/stoplightio/elements/issues/629)) ([d78b62d](https://github.com/stoplightio/elements/commit/d78b62d980c56f5572e0e264a46f05219018de4a))
+
+
+
+
+
+# [6.0.0-beta.169](https://github.com/stoplightio/elements/compare/v6.0.0-beta.168...v6.0.0-beta.169) (2020-10-16)
+
+
+### Features
+
+* allow to provide api documentation directly to the React API component ([#625](https://github.com/stoplightio/elements/issues/625)) ([869a67b](https://github.com/stoplightio/elements/commit/869a67bf8ff02d5fe3a1a246b81b3bc219b25658))
+
+
+
+
+
+# [6.0.0-beta.168](https://github.com/stoplightio/elements/compare/v6.0.0-beta.167...v6.0.0-beta.168) (2020-10-15)
+
+
+### Bug Fixes
+
+* breaking layout on long request param string ([#626](https://github.com/stoplightio/elements/issues/626)) ([aa99697](https://github.com/stoplightio/elements/commit/aa996971da29fa822f933aaecd4e8ddc532b72a0))
+
+
+
+
+
 # [6.0.0-beta.167](https://github.com/stoplightio/elements/compare/v6.0.0-beta.166...v6.0.0-beta.167) (2020-10-14)